+++ src/confload.nim
@@ -176,8 +176,6 @@ proc loadAllConfigs*() =
     addConfLoad(attestConfName,    toStream(attestConfig),    checkNone).
     addConfLoad(sbomConfName,      toStream(sbomConfig),      checkNone).
     addConfLoad(sastConfName,      toStream(sastConfig),      checkNone).
-    addConfLoad(techStackConfName, toStream(techStackConfig), checkNone).
-    addConfLoad(linguistConfName,  toStream(linguistConfig),  checkNone).
     addConfLoad(coConfName,        toStream(coConfig),        checkNone)
 
   stack.addCallback(loadLocalStructs)